context switch: round robin schedule policy achieved.

+++ os.h
@@ -5,16 +5,16 @@
 
 #define NO_OF_THREADS 2
 #define NO_OF_CPU_REGS 15
-#define STACK_SIZE    100 
+#define STACK_SIZE    100
 
 typedef struct s_thread t_thread;
 
 struct s_thread
 {
-    uint16_t  *sp;
+    uint32_t  *sp;
     t_thread *next;
     uint8_t   tid;
-    uint16_t  stack[STACK_SIZE];
+    uint32_t  stack[STACK_SIZE];
 };
 
 typedef void (*f_threadHandler)(void);
